refactor: task commands
