+++ .husky/commit-msg
@@ -1,4 +1,6 @@
 #!/bin/sh
 . "$(dirname "$0")/_/husky.sh"
 
+set -x
+
 yarn cli tasks commitlint --env test $1
