+++ apps/cli/src/commands/parse/parse.command.ts
@@ -6,10 +6,13 @@ import { __seedAtomData } from '../import/import-seed-data'
 
 export const parseCommand: CommandModule<any, any> = {
   command: 'parse',
-  describe: 'Parse Ant Design scraped files to Interface model',
+  describe:
+    'Parse Ant Design scraped CSV files and insert to application as types',
   handler: async () => {
     const { selectedUser } = await inquirer.prompt([await selectUserPrompt()])
 
     await __seedAtomData(selectedUser, antdAtomsFactory)
+
+    return process.exit(0)
   },
 }
