+++ apps/cli/src/commands/scrape/scrape.command.ts
@@ -3,7 +3,7 @@ import { ScraperService } from './scraper.service'
 
 export const scrapeCommand: CommandModule<unknown, unknown> = {
   command: 'scrape',
-  describe: 'Scrape props data from Ant Design',
+  describe: 'Scrape props data from Ant Design as CSV files',
   handler: async () => {
     const scraper = new ScraperService()
