+++ apps/cli/src/shared/utils/env.ts
@@ -19,8 +19,11 @@ export enum Stage {
   // Vercel remote
   Prod = 'prod',
 
-  // Vercel remote
-  Staging = 'staging',
+  // // Vercel remote
+  // Staging = 'staging',
+
+  // // Local test flow, exists in memory only
+  // Test = 'test',
 
   // Local
   Dev = 'dev',
